refactor: simplify LangSmith ingestion
